+++ src/main/java/com/bbots/gl/model/GlTransation.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class GlTransation {
 
-	private int orgCode;
+	private Long orgCode;
     private int glNo;
     private String allowedCurr;
     private String eUser;
